[ 1844941 ] Interpreter should cache exit routine addresses.
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@2139 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ platform/windows/orxscrpt/scrptdebug.hpp
@@ -93,7 +93,7 @@ typedef struct FLAGLISTstruct {
                                                a,                           \
                                                -1,                          \
                                                w,                           \
-                                               cb)
+                                               (int)(cb))
 
 //
 // WINBASEAPI
